Consolidate sql helpers
